Cria um servidor de teste único
