+++ sysarq/src/tests/Documents/Create/CreateAdministrativeProcess.test.js
@@ -1,151 +1,14 @@
-import { setupServer } from "msw/node";
-import { rest } from "msw";
-
 import { render, screen, fireEvent, within } from "@testing-library/react";
+
+import server from "../../support/testServer";
 
 import CreateAdministrativeProcess from "../../../pages/Documents/Create/CreateAdministrativeProcess";
 
-const hostApiArchives = `${process.env.REACT_APP_URL_API_ARCHIVES}`;
-
 jest.setTimeout(30000);
-
-const server = setupServer(
-	rest.get(`${hostApiArchives}document-subject/`, (req, res, ctx) => {
-		return res(
-			ctx.json([
-				{
-					id: 34,
-					subject_name: "subject_name_test",
-					temporality: "2035-09-10",
-				},
-			])
-		);
-	}),
-
-	rest.get(`${hostApiArchives}unity/`, (req, res, ctx) => {
-		return res(
-			ctx.json([
-				{
-					id: 38,
-					telephone_number: "39",
-					note: "unit_notes_test",
-					unity_name: "sender_unit_name_test",
-					unity_abbreviation: "unit_abbreviation_test",
-					administrative_bond: "unit_administrative_bond_test",
-					bond_abbreviation: "unit_bond_abbreviation_test",
-					type_of_unity: "type_of_unit_test",
-					municipality: "unit_municipality_test",
-				},
-				{
-					id: 40,
-					telephone_number: "41",
-					note: "unit_notes_test_1",
-					unity_name: "destination_unit_name_test",
-					unity_abbreviation: "unit_abbreviation_test_1",
-					administrative_bond: "unit_administrative_bond_test_1",
-					bond_abbreviation: "unit_bond_abbreviation_test_1",
-					type_of_unity: "type_of_unit_test_1",
-					municipality: "unit_municipality_test_1",
-				},
-				{
-					id: 41,
-					telephone_number: "42",
-					note: "unit_notes_test_2",
-					unity_name: "unarchive_unit_name_test",
-					unity_abbreviation: "unit_abbreviation_test_2",
-					administrative_bond: "unit_administrative_bond_test_2",
-					bond_abbreviation: "unit_bond_abbreviation_test_2",
-					type_of_unity: "type_of_unit_test_2",
-					municipality: "unit_municipality_test_2",
-				},
-			])
-		);
-	}),
-
-	rest.get(`${hostApiArchives}box-abbreviation/`, (req, res, ctx) => {
-		return res(
-			ctx.json([
-				{
-					id: 43,
-					number: 44,
-					abbreviation: "abbreviation_test",
-					name: "abbreviation_name_test",
-					year: 2045,
-				},
-			])
-		);
-	}),
-
-	rest.get(`${hostApiArchives}shelf/`, (req, res, ctx) => {
-		return res(
-			ctx.json([
-				{
-					id: 46,
-					number: 47,
-				},
-			])
-		);
-	}),
-
-	rest.get(`${hostApiArchives}rack/`, (req, res, ctx) => {
-		return res(
-			ctx.json([
-				{
-					id: 48,
-					number: 49,
-				},
-			])
-		);
-	}),
-
-	rest.post(`${hostApiArchives}administrative-process/`, (req, res, ctx) => {
-		if (
-			req.body.notice_date === "2005-04-03" &&
-			req.body.archiving_date === "2011-10-09" &&
-			req.body.reference_month_year === "2015-04-01" &&
-			req.body.process_number === "16" &&
-			req.body.cpf_cnpj === "28293031323" &&
-			req.body.interested === "interested_test" &&
-			req.body.subject_id === 34 &&
-			req.body.dest_unity_id === 40 &&
-			req.body.sender_unity === 38 &&
-			req.body.sender_user === "sender_worker_test" &&
-			req.body.abbreviation_id === 43 &&
-			req.body.shelf_id === 46 &&
-			req.body.rack_id === 48 &&
-			req.body.is_filed === false &&
-			req.body.is_eliminated === false &&
-			req.body.unity_id === 41 &&
-			req.body.administrative_process_number === "50" &&
-			req.body.send_date === "2055-09-08" &&
-			req.body.notes === "notes_test"
-		) {
-			return res(ctx.status(201));
-		} else {
-			return res(ctx.status(404));
-		}
-	})
-);
 
 beforeAll(() => server.listen());
 afterEach(() => server.resetHandlers());
 afterAll(() => server.close());
-
-const input = (field, value) => {
-	fireEvent.change(screen.getByLabelText(field), { target: { value: value } });
-};
-
-const submitClick = () => {
-	fireEvent.click(screen.getByRole("button", { name: /CADASTRAR/ }));
-};
-
-const isOnTheScreen = (text) => {
-	expect(screen.getByText(text)).toBeInTheDocument();
-};
-
-const isNotInTheDocument = (text) => {
-	expect(screen.queryByText(text)).not.toBeInTheDocument();
-};
 
 const NOTICE_DATE_LABEL = "Data de Autuação*";
 
@@ -161,6 +24,22 @@ const UNARCHIVE_PROCESS_NUMBER_LABEL = "Nº do Processo do Desarquivamento";
 
 const UNARCHIVE_DATE_LABEL = "Data de Desarquivamento";
 
+const input = (field, value) => {
+	fireEvent.change(screen.getByLabelText(field), { target: { value: value } });
+};
+
+const submitClick = () => {
+	fireEvent.click(screen.getByRole("button", { name: /CADASTRAR/ }));
+};
+
+const isOnTheScreen = (text) => {
+	expect(screen.getByText(text)).toBeInTheDocument();
+};
+
+const isNotOnTheScreen = (text) => {
+	expect(screen.queryByText(text)).not.toBeInTheDocument();
+};
+
 describe("Create Administrative Process Screen Test", () => {
 	it("complete test", async () => {
 		render(<CreateAdministrativeProcess />);
@@ -170,54 +49,54 @@ describe("Create Administrative Process Screen Test", () => {
 		isOnTheScreen(REQUIRED_DATE_ERROR_MESSAGE);
 
 		input(NOTICE_DATE_LABEL, "01/02/");
-		isNotInTheDocument(REQUIRED_DATE_ERROR_MESSAGE);
+		isNotOnTheScreen(REQUIRED_DATE_ERROR_MESSAGE);
 		submitClick();
 		isOnTheScreen(INVALID_DATE_ERROR_MESSAGE);
 
 		input(NOTICE_DATE_LABEL, "03/04/2005");
-		isNotInTheDocument(INVALID_DATE_ERROR_MESSAGE);
+		isNotOnTheScreen(INVALID_DATE_ERROR_MESSAGE);
 
 		input(ARCHIVING_DATE_LABEL, "");
 		submitClick();
 		isOnTheScreen(REQUIRED_DATE_ERROR_MESSAGE);
 
 		input(ARCHIVING_DATE_LABEL, "36/07/2008");
-		isNotInTheDocument(REQUIRED_DATE_ERROR_MESSAGE);
+		isNotOnTheScreen(REQUIRED_DATE_ERROR_MESSAGE);
 		submitClick();
 		isOnTheScreen(INVALID_DATE_ERROR_MESSAGE);
 
 		input(ARCHIVING_DATE_LABEL, "09/10/2011");
-		isNotInTheDocument(INVALID_DATE_ERROR_MESSAGE);
+		isNotOnTheScreen(INVALID_DATE_ERROR_MESSAGE);
 
 		input("Referência", "13/2012");
 		submitClick();
 		isOnTheScreen("Insira um período válido");
 
 		input("Referência", "04/2015");
-		isNotInTheDocument("Insira um período válido");
+		isNotOnTheScreen("Insira um período válido");
 
 		submitClick();
 		isOnTheScreen("Insira o número do processo");
 
 		input("Número do Processo*", "16");
-		isNotInTheDocument("Insira o número do processo");
+		isNotOnTheScreen("Insira o número do processo");
 
 		input("CPF/CNPJ", "171.819.20212");
 		submitClick();
 		isOnTheScreen("Insira somente números");
 
 		input("CPF/CNPJ", "2324252627");
-		isNotInTheDocument("Insira somente números");
+		isNotOnTheScreen("Insira somente números");
 		submitClick();
 		isOnTheScreen("Insira um CPF/CNPJ válido");
 
 		input("CPF/CNPJ", "28293031323");
-		isNotInTheDocument("Insira um CPF/CNPJ válido");
+		isNotOnTheScreen("Insira um CPF/CNPJ válido");
 		submitClick();
 		isOnTheScreen("Insira um interessado");
 
 		input("Interessado*", "interested_test");
-		isNotInTheDocument("Insira um interessado");
+		isNotOnTheScreen("Insira um interessado");
 		submitClick();
 		isOnTheScreen("Selecione um assunto");
 
@@ -225,7 +104,7 @@ describe("Create Administrative Process Screen Test", () => {
 		const subjectsOptions = within(screen.getByRole("listbox"));
 		await subjectsOptions.findByText("subject_name_test");
 		fireEvent.click(subjectsOptions.getByText(/subject_name_test/i));
-		isNotInTheDocument("Selecione um assunto");
+		isNotOnTheScreen("Selecione um assunto");
 
 		submitClick();
 		isOnTheScreen("Selecione uma unidade");
@@ -234,7 +113,7 @@ describe("Create Administrative Process Screen Test", () => {
 		const senderUnitOptions = within(screen.getByRole("listbox"));
 		await senderUnitOptions.findByText("sender_unit_name_test");
 		fireEvent.click(senderUnitOptions.getByText(/sender_unit_name_test/i));
-		isNotInTheDocument("Selecione uma unidade");
+		isNotOnTheScreen("Selecione uma unidade");
 
 		submitClick();
 		isOnTheScreen("Selecione um status");
@@ -242,7 +121,7 @@ describe("Create Administrative Process Screen Test", () => {
 		fireEvent.mouseDown(screen.getByLabelText("Status*"));
 		const statusOptions = within(screen.getByRole("listbox"));
 		fireEvent.click(statusOptions.getByText(/Eliminado/i));
-		isNotInTheDocument("Selecione um status");
+		isNotOnTheScreen("Selecione um status");
 
 		submitClick();
 
@@ -281,9 +160,9 @@ describe("Create Administrative Process Screen Test", () => {
 		await rackOptions.findByText("49");
 		fireEvent.click(rackOptions.getByText(/49/i));
 
-		isNotInTheDocument(UNARCHIVE_DESTINATION_UNIT_LABEL);
-		isNotInTheDocument(UNARCHIVE_PROCESS_NUMBER_LABEL);
-		isNotInTheDocument(UNARCHIVE_DATE_LABEL);
+		isNotOnTheScreen(UNARCHIVE_DESTINATION_UNIT_LABEL);
+		isNotOnTheScreen(UNARCHIVE_PROCESS_NUMBER_LABEL);
+		isNotOnTheScreen(UNARCHIVE_DATE_LABEL);
 
 		fireEvent.mouseDown(screen.getByLabelText("Status*"));
 		const statusOptions2 = within(screen.getByRole("listbox"));
@@ -313,7 +192,7 @@ describe("Create Administrative Process Screen Test", () => {
 		isOnTheScreen(INVALID_DATE_ERROR_MESSAGE);
 
 		input(UNARCHIVE_DATE_LABEL, "08/09/2055");
-		isNotInTheDocument(INVALID_DATE_ERROR_MESSAGE);
+		isNotOnTheScreen(INVALID_DATE_ERROR_MESSAGE);
 
 		input("Observação", "notes_test");
